+++ docs/conf.py
@@ -42,7 +42,7 @@ copyright = "2012 - 2021 Michal Čihař"
 author = "Michal Čihař"
 
 # The full version, including alpha/beta/rc tags
-release = "4.7.1"
+release = "4.8"
 
 
 # -- General configuration ---------------------------------------------------